requirements.txt
gTTS
SpeechRecognition
python-dotenv
langchain-ollama
pyaudio